refactor(plugins): vertical-slice agent-id-auth into its plugin dir
Third per-plugin vertical slice. The largest extraction by line count
— the service-manifest discovery+validation block (~570 lines) moves
to plugins/agent-id-auth/lib/manifest.mjs. Plugin-private: only this
plugin needs to parse and render manifests, so it does not live in core.

- plugins/agent-id-auth/lib/manifest.mjs — parseServiceManifest,
  renderCapabilities, fetchServiceManifest, buildServiceAuthHeader,
  probeServiceSupportSignal, resolveServiceApiUrl, plus the
  ~30 closed-key constants and the recursive operation/schema/property
  validators they rely on. Pure validation surface; no other dependencies.
- plugins/agent-id-auth/bin/cli.mjs — focused auth CLI with header,
  call, discover, capabilities, support subcommands (renamed from
  auth-header / discover-service / service-support now that the
  prefixes are redundant in a dedicated binary).
- plugins/agent-id-auth/skills/agent-id-auth/SKILL.md — surfaces the
  plugin to Claude Code with usage, trust-boundary rules, and the
  two-header pattern for cases where the agent must drive curl itself.

With this commit skills/alien-agent-id/lib.mjs is down to 96 lines —
purely re-exports + internal imports for the legacy meta-CLI. All five
core modules and three plugin libs are addressable from their own
directories. Tests: 184/184 pass.
